# [prototype] Optimize and clean up all affine methods