--- a/torchvision/prototype/transforms/functional/_geometry.py
+++ b/torchvision/prototype/transforms/functional/_geometry.py
@@ -1036,13 +1036,14 @@ def crop_image_tensor(image: torch.Tensor, top: int, left: int, height: int, wid
     bottom = top + height
 
     if left < 0 or top < 0 or right > w or bottom > h:
-        padding_ltrb = [
+        image = image[..., max(top, 0) : bottom, max(left, 0) : right]
+        torch_padding = [
             max(min(right, 0) - left, 0),
-            max(min(bottom, 0) - top, 0),
             max(right - max(w, left), 0),
+            max(min(bottom, 0) - top, 0),
             max(bottom - max(h, top), 0),
         ]
-        return pad_image_tensor(image[..., max(top, 0) : bottom, max(left, 0) : right], padding_ltrb)
+        return _pad_with_scalar_fill(image, torch_padding, fill=0, padding_mode="constant")
     return image[..., top:bottom, left:right]