@datumbox 2022-11-11
There are some related [failures](https://github.com/pytorch/vision/actions/runs/3447323984/jobs/5753214048). I'll check them on Monday.