@datumbox 2022-11-14
@pmeier We observe a speed improvement. The [flaky](https://github.com/pytorch/vision/actions/runs/3462890917/jobs/5782376784) test at prototype is a known issue on the bboxes and Victor is looking at it. Merging.